+++ cli/cli/commands/service/update/update.go
@@ -13,7 +13,7 @@ import (
 	"github.com/kurtosis-tech/kurtosis/cli/cli/command_framework/lowlevel/args"
 	"github.com/kurtosis-tech/kurtosis/cli/cli/command_framework/lowlevel/flags"
 	"github.com/kurtosis-tech/kurtosis/cli/cli/command_str_consts"
-	"github.com/kurtosis-tech/kurtosis/cli/cli/commands/service"
+	"github.com/kurtosis-tech/kurtosis/cli/cli/commands/service/service_helpers"
 	"github.com/kurtosis-tech/kurtosis/cli/cli/out"
 	"github.com/kurtosis-tech/kurtosis/container-engine-lib/lib/backend_interface"
 	"github.com/kurtosis-tech/kurtosis/metrics-library/golang/lib/metrics_client"
@@ -84,20 +84,20 @@ var ServiceUpdateCmd = &engine_consuming_kurtosis_command.EngineConsumingKurtosi
 	EngineClientContextKey:    engineClientCtxKey,
 	Flags: []*flags.FlagConfig{
 		{
-			Key:     service.ImageKey,
+			Key:     service_helpers.ImageKey,
 			Usage:   "image",
 			Type:    flags.FlagType_String,
 			Default: "",
 		},
 		{
-			Key:   service.CmdKey,
+			Key:   service_helpers.CmdKey,
 			Usage: "cmd",
 			// TODO Make this a string list
 			Type:    flags.FlagType_String,
 			Default: "",
 		},
 		{
-			Key:   service.EntrypointFlagKey,
+			Key:   service_helpers.EntrypointFlagKey,
 			Usage: "ENTRYPOINT binary that will be used when running the container, overriding the image's default ENTRYPOINT",
 			// TODO Make this a string list
 			Type:    flags.FlagType_String,
@@ -105,7 +105,7 @@ var ServiceUpdateCmd = &engine_consuming_kurtosis_command.EngineConsumingKurtosi
 		},
 		{
 			// TODO We currently can't handle commas, so allow users to set the flag multiple times to set multiple envvars
-			Key: service.EnvvarsFlagKey,
+			Key: service_helpers.EnvvarsFlagKey,
 			Usage: fmt.Sprintf(
 				"String containing environment variables that will be set when running the container, in "+
 					"the form \"KEY1%vVALUE1%vKEY2%vVALUE2\"",
@@ -117,7 +117,7 @@ var ServiceUpdateCmd = &engine_consuming_kurtosis_command.EngineConsumingKurtosi
 			Default: "",
 		},
 		{
-			Key: service.PortsFlagKey,
+			Key: service_helpers.PortsFlagKey,
 			Usage: fmt.Sprintf(`String containing declarations of ports that the container will listen on, in the form, %q`+
 				` where %q is a user friendly string for identifying the port, %q is required field, %q is an optional field which must be either`+
 				` '%v' or '%v' and defaults to '%v' if omitted and %q is user defined optional value. %v`,
@@ -135,7 +135,7 @@ var ServiceUpdateCmd = &engine_consuming_kurtosis_command.EngineConsumingKurtosi
 			Default: "",
 		},
 		{
-			Key: service.FilesFlagKey,
+			Key: service_helpers.FilesFlagKey,
 			Usage: fmt.Sprintf(
 				"String containing declarations of files paths on the container -> artifact name  where the contents of those "+
 					"files artifacts should be mounted, in the form \"MOUNTPATH1%vARTIFACTNAME1%vMOUNTPATH2%vARTIFACTNAME2\" where "+
@@ -230,7 +230,7 @@ func run(
 		return stacktrace.Propagate(err, "An error occurred getting the env vars string using key '%v'", envvarsFlagKey)
 	}
 	if envVarsStr != "" {
-		overrideEnvVars, err = service.ParseEnvVarsStr(envVarsStr)
+		overrideEnvVars, err = service_helpers.ParseEnvVarsStr(envVarsStr)
 		if err != nil {
 			return stacktrace.Propagate(err, "An error occurred parsing env vars string: %v", envVarsStr)
 		}
@@ -241,7 +241,7 @@ func run(
 		return stacktrace.Propagate(err, "An error occurred getting the ports string using key '%v'", portsFlagKey)
 	}
 	if portsStr != "" {
-		overridePorts, err = service.ParsePortsStr(portsStr)
+		overridePorts, err = service_helpers.ParsePortsStr(portsStr)
 		if err != nil {
 			return stacktrace.Propagate(err, "An error occurred parsing ports string: %v", portsStr)
 		}
@@ -252,13 +252,13 @@ func run(
 		return stacktrace.Propagate(err, "An error occurred getting the files artifact mounts string using key '%v'", filesFlagKey)
 	}
 	if filesArtifactMountsStr != "" {
-		overrideFilesArtifactsMountpoint, err = service.ParseFilesArtifactMountsStr(filesArtifactMountsStr)
+		overrideFilesArtifactsMountpoint, err = service_helpers.ParseFilesArtifactMountsStr(filesArtifactMountsStr)
 		if err != nil {
 			return stacktrace.Propagate(err, "An error occurred parsing files artifacts mount points string: %v", filesArtifactMountsStr)
 		}
 	}
 
-	_, currServiceConfig, err := service.GetServiceInfo(ctx, kurtosisCtx, enclaveIdentifier, serviceName)
+	_, currServiceConfig, err := service_helpers.GetServiceInfo(ctx, kurtosisCtx, enclaveIdentifier, serviceName)
 	if err != nil {
 		return stacktrace.Propagate(err, "An error occurred getting service info of service '%v' in enclave '%v'.", serviceName, enclaveIdentifier)
 	}
@@ -343,10 +343,10 @@ func run(
 	)
 	//logrus.Infof("SERVICE CONFIG STRING: %v", serviceConfigStr)
 
-	addServiceStarlarkStr := service.GetAddServiceStarlarkScript(serviceName, serviceConfigStr)
+	addServiceStarlarkStr := service_helpers.GetAddServiceStarlarkScript(serviceName, serviceConfigStr)
 
 	logrus.Infof("Running update service starlark for service '%v' in enclave '%v'...", serviceName, enclaveIdentifier)
-	starlarkRunResult, err := service.RunAddServiceStarlarkScript(ctx, serviceName, enclaveIdentifier, addServiceStarlarkStr, enclaveCtx)
+	starlarkRunResult, err := service_helpers.RunAddServiceStarlarkScript(ctx, serviceName, enclaveIdentifier, addServiceStarlarkStr, enclaveCtx)
 	if err != nil {
 		return err //already wrapped
 	}
